+++ covalent_awslambda_plugin/awslambda.py
@@ -249,7 +249,7 @@ class AWSLambdaExecutor(AWSExecutor):
                 if lambda_state["Configuration"]["State"] == "Active":
                     is_active = True
                 else:
-                    asyncio.sleep(0.5)
+                    await asyncio.sleep(0.5)
                     continue
         return is_active
 
@@ -332,15 +332,17 @@ class AWSLambdaExecutor(AWSExecutor):
 
         with self.get_session() as session:
             while not self._key_exists:
-                asyncio.sleep(0.5)
                 s3_client = session.client("s3")
-
                 try:
                     current_keys = [
                         item["Key"]
                         for item in s3_client.list_objects(Bucket=self.s3_bucket_name)["Contents"]
                     ]
                     self._key_exists = object_key in current_keys
+
+                    if not self._key_exists:
+                        await asyncio.sleep(0.5)
+
                     return self._key_exists
                 except botocore.exceptions.ClientError as ce:
                     app_log.exception(ce)